feat - implement login and sign-up with saved id
Signing up now checks whether the username is unique, and if it is, adds the details into the database and logs the user in. Logging in now checks whether the username/password combination is correct before logging in. The user's ID is temporarily being saved in the controller as a public field.