feat - check history if the word has been played before the reset
The programme now keeps a running history of games played. It will repeat a word in a category according to a user's individual history UNTIL it has played through all of the words and there are no more words left. Then, the history will clear, and all words for that category is free game again. This will be specific to each user.